Added networking for object entities
Added networking for object entities
Casted cof component in flatbuffers to ensure they are sign extended properly (0xFF was not becoming 0xFFFFFFFF which is a special value)
Added additional check in CofManager#setComponent to ignore change if redundant
Increased output packet queue to 1024 -- will eventually group all sync packets into a single packet

+++ core/src/com/riiablo/engine/server/SerializationManager.java
@@ -151,7 +151,7 @@ public class SerializationManager extends PassiveSystem {
           CofComponentsSerializer serializer = (CofComponentsSerializer) serializers.get(clazz);
           CofComponentsP table = serializer.getTable(sync, i);
           for (int j = 0, s = table.componentLength(); j < s; j++) {
-            cofs.setComponent(entityId, j, table.component(j));
+            cofs.setComponent(entityId, j, (byte) table.component(j));
           }
 //          if (DEBUG_DESERIALIZE) Gdx.app.log(TAG, "  " + DebugUtils.toByteArray(ArrayUtils.toByteArray(mCofComponents.get(entityId).component)));
           break;
